+++ x/leverage/keeper/token.go
@@ -5,6 +5,31 @@ import (
 	"github.com/umee-network/umee/v4/x/leverage/types"
 )
 
+// CleanTokenRegistry deletes all blacklisted tokens in the leverage registry
+// whose uToken supplies are zero. Called automatically on registry update.
+func (k Keeper) CleanTokenRegistry(ctx sdk.Context) error {
+	tokens := k.GetAllRegisteredTokens(ctx)
+	for _, t := range tokens {
+		if t.Blacklist {
+			uDenom := types.ToUTokenDenom(t.BaseDenom)
+			uSupply := k.GetUTokenSupply(ctx, uDenom)
+			if uSupply.IsZero() {
+				k.deleteTokenSettings(ctx, t.BaseDenom)
+			}
+		}
+	}
+	return nil
+}
+
+// deleteTokenSettings deletes a Token in the x/leverage module's KVStore.
+// it should only be called by CleanTokenRegistry.
+func (k Keeper) deleteTokenSettings(ctx sdk.Context, denom string) error {
+	store := ctx.KVStore(k.storeKey)
+	tokenKey := types.KeyRegisteredToken(denom)
+	store.Delete(tokenKey)
+	return nil
+}
+
 // SetTokenSettings stores a Token into the x/leverage module's KVStore.
 func (k Keeper) SetTokenSettings(ctx sdk.Context, token types.Token) error {
 	if err := token.Validate(); err != nil {